+++ src/dlls/objects/334_CFPrisonGua/CFPrisonGua.c
@@ -2,13 +2,20 @@
 
 #include "dlls/objects/334_CFPrisonGua.h"
 
+#include "main/audio/sfx_play_api.h"
+#include "main/audio/sfx_stop_channel_api.h"
 #include "main/audio/sfx_trigger_ids.h"
 #include "main/dll/objfx_api.h"
+#include "main/dll/player_api.h"
 #include "main/dll/waterfx.h"
 #include "main/frame_timing.h"
 #include "main/gamebits.h"
+#include "main/obj_message.h"
 #include "main/object_render.h"
+#include "main/object_update_list.h"
+#include "main/objhits.h"
 #include "main/objseq.h"
+#include "main/render_lactions_api.h"
 #include "main/vecmath.h"
 #include "sys/objects.h"
 
@@ -151,9 +158,9 @@ int cfPrisonGuard_sequenceCallback(GameObject* obj, int unused, ObjSeqState* ani
         break;
     }
     if (obj->anim.currentMove == CFPRISONGUARD_WATER_MOVE_B || obj->anim.currentMove == CFPRISONGUARD_WATER_MOVE_A) {
-        Sfx_PlayFromObject((int)obj, SFXTRIG_foot_water_roll);
+        Sfx_PlayFromObject(obj, SFXTRIG_foot_water_roll);
     } else {
-        Sfx_StopObjectChannel((int)obj, CFPRISONGUARD_WATER_SFX_CHANNEL);
+        Sfx_StopObjectChannel(obj, CFPRISONGUARD_WATER_SFX_CHANNEL);
     }
     if (uncleFlewOff != 0 && state->uncleFlewOffLatch == 0) {
         shouldTransition = 1;
