+++ src/dlls/objects/395_CClevcontro/CClevcontro.c
@@ -7,12 +7,16 @@
  */
 #include "dlls/objects/395_CClevcontro.h"
 
+#include "main/audio/music_api.h"
 #include "main/audio/music_trigger_ids.h"
+#include "main/audio/sfx_play_api.h"
 #include "main/audio/sfx_trigger_ids.h"
 #include "main/camera_interface.h"
+#include "main/dll/savegame_load_api.h"
 #include "main/frame_timing.h"
 #include "main/gamebit_ids.h"
 #include "main/gamebits_api.h"
+#include "main/gametext_show_api.h"
 #include "main/mapEventTypes.h"
 #include "main/object_render.h"
 #include "main/objfx.h"
@@ -181,7 +185,7 @@ void ccLevelControl_update(GameObject* obj) {
     heldGoldBarCount = mainGetBit(GAMEBIT_ITEM_CCGoldBar_Count);
     if (heldGoldBarCount + usedGoldBarCount == CC_LEVEL_CONTROL_GOLD_BAR_TARGET &&
         mainGetBit(CC_LEVEL_CONTROL_GOLD_BAR_COMPLETION_SFX_GAMEBIT) == 0) {
-        Sfx_PlayFromObject((int)obj, SFXTRIG_mpick1_b);
+        Sfx_PlayFromObject(obj, SFXTRIG_mpick1_b);
         mainSetBits(CC_LEVEL_CONTROL_GOLD_BAR_COMPLETION_SFX_GAMEBIT, 1);
     }
 }
